Create sync_all_media.sh
The script starts out using either yad or zenity to ask the user to select a drive in the /media directory (must be in the /media directory or there is an error and the script exits) and then looks at the remaining peripherals and compares the used space on the host drive with the total space on the peripheral and only shows those that are larger than the host enough to hold a one-to-one backup on the destination drive.

It also removes any listing of a Windows OS that may be loaded in the /media directory and if found, unmounts the Windows OS partition before starting the drive survey.

The script uses df to get the drive information for each drive and each column of information is loaded into an array for each drive, sorted by the filesystem path name, alphabetically. The information is the drive name, drive path, drive filesystem path, total space, used space, and available space, to see all size information.

The script allows you to choose one or more media drives to run the rsync program on and will run the sync in the sequence you provide, so if five USB or backup peripherals were found and you want to do the back up in the order of "5 2 1 4 3", it will do just that for each selection. Selections also for syncing media host to all media drives or or a single directory to a media drive. A little animation spins while rsync is gathering file information until rsync starts.