Update sync_all_media.sh
Almost a complete rewrite. There were several issues fixed including many syntax errors. Fixed a problem with arrays and slimmed down the code. Had a problem with running the "Working....." animation and I think I fixed the problem with making sure the animation ceases and allows the return to the menu after completing the rsync process. Set it up so that it is working within a while/do statement with the rsync program as the determining factor. Seems like it works well at the moment but need more testers. This is still in the alpha stage but ready to shift into beta.