Update sync_all_media.sh
The Working animation was not receiving the final '0' return to stop its run. Changed to looking for the PID of the program's background run and check for finish when the PID is gone and wait for a return code for the return of the exit code. 

Seems to be in working order and no issues have come up so far in testing.