Update sync_all_media.sh
Fixed problem of multiple choices hanging because of entry of spaces or duplicating characters. Now it filters for these and converts all options in the variables being used to uppercase so there is no problem with defining the options being used. All the variables are placed into an array to allow for easy transition to the next drive sync run.