pause.sh
This is a simple pause script that interrupts the current process indefinitely until any key is pressed, or until the optional timer reaches 00. 
Script allows for entry of a timer (-t, --timer) in seconds, custom text for the prompt (-p "text", --prompt "text") and custom text for the 
response (-r "text", --response "text") after continuing.